requirements.txt
Django==4.1.2
djangorestframework==3.14.0
psycopg2-binary==2.9.4
django-cors-headers==3.13.0
requests